move Exasol tests
