+++ tests/Functional/Exasol/Schema/ExasolSchemaQueryBuilderTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Tests\Keboola\TableBackendUtils\Functional\Schema\Exasol;
+namespace Tests\Keboola\TableBackendUtils\Functional\Exasol\Schema;
 
 use Doctrine\DBAL\Exception;
 use Keboola\TableBackendUtils\DataHelper;
